Move file fetching methods to files.go
